Rename and move Authentication to AuthenticationService
